# Optimize by reducing heap allocations